test: classify managed v2 process contract
