+++ packages/testkit/src/operation-recovery-process.test.ts
@@ -136,90 +136,6 @@ test("operation recovery is single-owner and resumes with a higher attempt after
   }
 }, 20_000);
 
-test("managed v2 wait crash truth and the resumed real operation deadline survive process boundaries", async () => {
-  const testRoot = await mkdtemp(join(tmpdir(), "adam-agent-managed-v2-process-"));
-  const packageRoot = join(testRoot, "extension");
-  const stateRoot = join(testRoot, "state");
-  const workspaceRoot = join(testRoot, "workspace");
-  const deadlineStateRoot = join(testRoot, "deadline-state");
-  const deadlineWorkspaceRoot = join(testRoot, "deadline-workspace");
-  await mkdir(workspaceRoot);
-  await mkdir(deadlineWorkspaceRoot);
-  await writeManagedV2ProcessExtension(packageRoot);
-  const waitingOwner = spawnFixture({
-    mode: "managed-wait-start",
-    packageRoot,
-    stateRoot,
-    workspaceRoot,
-  });
-  let queryProcess: ChildProcess | undefined;
-  let deadlineProcess: ChildProcess | undefined;
-
-  try {
-    const waiting = await waitForMessage(waitingOwner, "managed-wait-durable");
-    const operationId = messageString(waiting, "operationId");
-    await waitForMessage(waitingOwner, "managed-model-started");
-    waitingOwner.kill("SIGKILL");
-    await waitForClose(waitingOwner);
-    const crashedRecords = await createJsonlOperationStore({ stateRoot, workspaceRoot }).then(
-      (store) => store.read(operationId),
-    );
-    expect(crashedRecords.map((record) => record.event.type)).toEqual([
-      "operation_started",
-      "operation_artifact_published",
-      "operation_managed_wait_started",
-    ]);
-
-    queryProcess = spawnFixture({
-      mode: "managed-query",
-      operationId,
-      packageRoot,
-      stateRoot,
-      workspaceRoot,
-    });
-    const queried = await waitForMessage(queryProcess, "managed-query-completed");
-    expect(queried).toMatchObject({
-      eventTypes: [
-        "operation_started",
-        "operation_artifact_published",
-        "operation_managed_wait_started",
-      ],
-      snapshot: { operationId, status: "recovery_required" },
-    });
-    await waitForClose(queryProcess);
-
-    deadlineProcess = spawnFixture({
-      mode: "managed-deadline",
-      packageRoot,
-      stateRoot: deadlineStateRoot,
-      workspaceRoot: deadlineWorkspaceRoot,
-    });
-    const deadline = await waitForMessage(deadlineProcess, "managed-deadline-terminal");
-    expect(deadline).toMatchObject({
-      eventTypes: [
-        "operation_started",
-        "operation_artifact_published",
-        "operation_managed_wait_started",
-        "operation_managed_wait_settled",
-        "operation_failed",
-      ],
-      snapshot: {
-        error: { code: "operation_deadline_exceeded" },
-        status: "failed",
-      },
-    });
-    await waitForClose(deadlineProcess);
-  } finally {
-    for (const child of [waitingOwner, queryProcess, deadlineProcess]) {
-      if (child !== undefined && child.exitCode === null && child.signalCode === null) {
-        child.kill("SIGKILL");
-        await waitForClose(child);
-      }
-    }
-    await rm(testRoot, { recursive: true, force: true });
-  }
-}, 20_000);
-
 function spawnFixture(options: {
   readonly mode: string;
   readonly operationId?: string;
@@ -390,66 +306,6 @@ async function writeProcessRecoveryExtension(packageRoot: string): Promise<void>
         output: JSON.parse(new TextDecoder().decode(bytes)),
         artifacts: [record.value.artifact],
       };
-    },
-  });
-}
-`,
-    "utf8",
-  );
-}
-
-async function writeManagedV2ProcessExtension(packageRoot: string): Promise<void> {
-  await mkdir(packageRoot, { recursive: true });
-  await writeFile(
-    join(packageRoot, "package.json"),
-    JSON.stringify({
-      name: "@fixture/managed-v2-process-extension",
-      version: "2.0.0",
-      type: "module",
-      adamAgent: {
-        id: "fixture.extension",
-        apiVersion: "^0.5.0",
-        runtime: { entry: "./runtime.js" },
-        capabilities: {
-          required: [
-            { id: "adam.artifact.publish@1", version: "^1.0.0" },
-            { id: "adam.managed-session@2", version: "^2.0.0" },
-          ],
-          optional: [],
-        },
-        contributions: [
-          {
-            kind: "operation",
-            id: "fixture.managed-review-v2",
-            input: { id: "fixture.input", version: 1 },
-            output: { id: "fixture.output", version: 1 },
-            progress: { id: "fixture.progress", version: 1 },
-            managedOutput: { id: "fixture.managed-output", version: 1 },
-          },
-        ],
-      },
-    }),
-    "utf8",
-  );
-  await writeFile(
-    join(packageRoot, "runtime.js"),
-    `export function activate(context) {
-  const codec = (id) => ({ id, version: 1, decode(value) { return { ok: true, value }; }, encode(value) { return { ok: true, value }; } });
-  const managedOutput = { id: "fixture.managed-output", version: 1, decode(value) { return value?.verdict === "verified" ? { ok: true, value } : { ok: false, issues: [] }; }, encode(value) { return this.decode(value); } };
-  context.registerOperation({
-    id: "fixture.managed-review-v2",
-    input: codec("fixture.input"), output: codec("fixture.output"), progress: codec("fixture.progress"), managedOutput,
-    async execute(_input, operation) {
-      const evidence = await operation.capabilities["adam.artifact.publish@1"].publish({ bytes: new TextEncoder().encode("immutable review evidence"), contract: { id: "fixture.review-evidence", version: 1 }, mediaType: "text/plain; charset=utf-8" });
-      const terminal = await operation.capabilities["adam.managed-session@2"].run({ evidence: [{ type: "artifact", artifact: evidence }], managedRole: "Review immutable evidence.", output: { id: "fixture.managed-output", version: 1 }, profile: { id: "reviewer.v1", version: 1 }, selectedSkills: [], task: "Return candidates." });
-      if (process.env.ADAM_AGENT_OPERATION_FIXTURE_MODE === "managed-deadline") {
-        await new Promise((resolve, reject) => {
-          const rejectAbort = () => reject(operation.signal.reason);
-          if (operation.signal.aborted) rejectAbort();
-          else operation.signal.addEventListener("abort", rejectAbort, { once: true });
-        });
-      }
-      return { terminal };
     },
   });
 }
